+++ Enlighter.php
@@ -3,8 +3,8 @@
 	Plugin Name: Enlighter - Customizable Syntax Highlighter
 	Plugin URI: http://enlighterjs.org
 	Description: Enlighter is a free, easy-to-use, syntax highlighting tool with a build-in theme editor.
-	Version: 2.12-BETA5
-	Author: Andi Dittrich
+	Version: 3.0-BETA1
+    Author: Andi Dittrich
 	Author URI: http://andidittrich.de
 	License: MIT X11-License
 	
@@ -22,7 +22,7 @@
 */
 
 define('ENLIGHTER_INIT', true);
-define('ENLIGHTER_VERSION', '2.12-BETA5');
+define('ENLIGHTER_VERSION', '3.0-BETA1');
 define('ENLIGHTER_PLUGIN_PATH', dirname(__FILE__));
 
 // check php version