+++ class/ConfigGenerator.php
@@ -1,13 +1,13 @@
 <?php
 /**
-	Enlighter JS COnfig Generator
-	Version: 1.0
+	Enlighter Javascript Config Generator
+	Version: 1.2
 	Author: Andi Dittrich
 	Author URI: http://andidittrich.de
 	Plugin URI: http://andidittrich.de/go/enlighterjs
 	License: MIT X11-License
 	
-	Copyright (c) 2013-2014, Andi Dittrich
+	Copyright (c) 2013-2016, Andi Dittrich
 
 	Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
 	
@@ -20,19 +20,20 @@ namespace Enlighter;
 class ConfigGenerator{
 	
 	private $_config;
-	private $_cacheFile = '';
+	private $_cacheFilename = 'EnlighterJS.init.js';
+    private $_cacheManager = null;
 	
 	public function __construct($settingsUtil, $cacheManager){
 		$this->_config = $settingsUtil->getOptions();
-		$this->_cacheFile = $cacheManager->getCachePath().'EnlighterJS.init.js';
+        $this->_cacheManager = $cacheManager;
 	}
 	
 	public function isCached(){
-		return file_exists($this->_cacheFile);
+        return $this->_cacheManager->fileExists($this->_cacheFilename);
 	}
 	
 	// generate js based config
-	public function getJSConfig(){
+	public function getInitializationConfig(){
 		$c = 'window.addEvent(\'domready\', function(){';
 		$c .= 'if (typeof EnlighterJS == "undefined"){return;};';
 		$c .= 'EnlighterJS.Util.Init(\''. $this->_config['selector'] .'\', \''. $this->_config['selectorInline'] .'\', ';
@@ -59,6 +60,26 @@ class ConfigGenerator{
 	
 	// store generated config into cachefile
 	public function generate(){
-		file_put_contents($this->_cacheFile, $this->getJSConfig());
+        $this->_cacheManager->writeFile($this->_cacheFilename, $this->getInitializationConfig());
 	}
+
+    public function getEditorPluginConfig(){
+        $c = 'Enlighter_EditorConfig = ';
+
+        // create config object
+        $c .= json_encode(array(
+            'languages' => \Enlighter::getAvailableLanguages(),
+            'themes' => \Enlighter::getAvailableThemes(),
+            'config' => array(
+                'theme' => $this->_config['defaultTheme'],
+                'language' => $this->_config['defaultLanguage'],
+                'linenumbers' => ($this->_config['linenumbers'] ? true : false),
+                'indent' => intval($this->_config['indent']),
+                'quicktagMode' => $this->_config['editorQuicktagMode'],
+                'languageShortcode' => ($this->_config['languageShortcode'] ? true : false)
+            )
+        ));
+
+        return $c;
+    }
 }